refactor(ALL-1461): extract preview generation logic into a custom hook
